refactor: Move port and exchange collectors to generic
