+++ src/capellambse_context_diagrams/collectors/generic.py
@@ -14,11 +14,17 @@ import logging
 import typing as t
 
 import capellambse.model as m
-from capellambse.metamodel import interaction, la, oa, pa, sa
+from capellambse.metamodel import cs, fa, interaction, la, oa, pa, sa
 from capellambse.model import DiagramType as DT
 
 from .. import _elkjs, context, filters
 from . import makers
+
+if t.TYPE_CHECKING:
+    Filter: t.TypeAlias = cabc.Callable[
+        [cabc.Iterable[m.ModelElement]],
+        cabc.Iterable[m.ModelElement],
+    ]
 
 logger = logging.getLogger(__name__)
 
@@ -314,3 +320,71 @@ def make_owner_boxes(
             current, make_box_func, boxes, boxes_to_delete
         )
     return current.uuid
+
+
+def port_collector(
+    target: m.ModelElement | m.ElementList, diagram_type: DT
+) -> tuple[dict[str, m.ModelElement], dict[str, m.ModelElement]]:
+    """Collect ports from `target` savely."""
+
+    def __collect(target):
+        port_types = fa.FunctionPort | fa.ComponentPort | cs.PhysicalPort
+        incoming_ports: dict[str, m.ModelElement] = {}
+        outgoing_ports: dict[str, m.ModelElement] = {}
+        for attr in DIAGRAM_TYPE_TO_CONNECTOR_NAMES[diagram_type]:
+            try:
+                ports = getattr(target, attr)
+                if not ports or not isinstance(ports[0], port_types):
+                    continue
+
+                if attr == "inputs":
+                    incoming_ports.update({p.uuid: p for p in ports})
+                elif attr == "ports":
+                    for port in ports:
+                        if port.direction == "IN":
+                            incoming_ports[port.uuid] = port
+                        else:
+                            outgoing_ports[port.uuid] = port
+                else:
+                    outgoing_ports.update({p.uuid: p for p in ports})
+            except AttributeError:
+                pass
+        return incoming_ports, outgoing_ports
+
+    if isinstance(target, cabc.Iterable):
+        assert not isinstance(target, m.ModelElement)
+        incoming_ports: dict[str, m.ModelElement] = {}
+        outgoing_ports: dict[str, m.ModelElement] = {}
+        for obj in target:
+            inc, out = __collect(obj)
+            incoming_ports.update(inc)
+            outgoing_ports.update(out)
+    else:
+        incoming_ports, outgoing_ports = __collect(target)
+    return incoming_ports, outgoing_ports
+
+
+def _extract_edges(
+    obj: m.ModelElement,
+    attribute: str,
+    filter: Filter,
+) -> cabc.Iterable[m.ModelElement]:
+    return filter(getattr(obj, attribute, []))
+
+
+def port_exchange_collector(
+    ports: t.Iterable[m.ModelElement],
+    filter: Filter = lambda i: i,
+) -> dict[str, list[fa.AbstractExchange]]:
+    """Collect exchanges from `ports` savely."""
+    edges: dict[str, list[fa.AbstractExchange]] = {}
+
+    for port in ports:
+        if exs := _extract_edges(port, "exchanges", filter):
+            edges[port.uuid] = t.cast(list[fa.AbstractExchange], exs)
+            continue
+
+        if links := _extract_edges(port, "links", filter):
+            edges[port.uuid] = t.cast(list[fa.AbstractExchange], links)
+
+    return edges
